Decompose a time series object

Decompose a time series object

decomp(
  x,
  var = NULL,
  sm_model = c("stl"),
  mult_fit = FALSE,
  model_range = NULL,
  ...
)

# S3 method for inz_dcmp
plot(
  x,
  recompose.progress = c(0, 0),
  recompose = any(recompose.progress > 0),
  ylab = NULL,
  title = NULL,
  colour = c("#1B9E46", "#45a8ff", "orangered"),
  ...
)

Arguments

x

a decomp (inz_dcmp) object

var

a character vector of length one, or NULL

sm_model

the smoothing method to be used

mult_fit

If TRUE, a multiplicative model is used, otherwise an additive model is used by default.

model_range

range of data to be decomposed by the model, specified as dates or years, if part of model_range specified is outside the range of the data, the exceeding proportion is ignored.

...

additional arguments (ignored)

recompose.progress

if recompose is TRUE, this shows how much to show (for animation!). Length 2 numeric: the first is 0 for seasonal, and 1 for residual; second component is how many observations have been recomposed so far

recompose

logical as to whether the recomposition is shown or not

ylab

the label for the y axis

title

the title for the plot

colour

vector of three colours for trend, seasonal, and residuals, respectively

Value

a decomp (inz_dcmp) object, a sub-class of dable

References

R. B. Cleveland, W. S. Cleveland, J.E. McRae, and I. Terpenning (1990) STL: A Seasonal-Trend Decomposition Procedure Based on Loess. Journal of Official Statistics, 6, 3iV73.

See also

Examples

ts <- inzightts(visitorsQ)
d <- decomp(ts)
#> Decomp variable not specified, automatically selected `var = "Australia"`

if (FALSE) {
plot(d)
}