Summarise iNZightTS forecasts

Summary method for objects of class inz_frct.

# S3 method for inz_frct
summary(object, var = NULL, ...)

# S3 method for summary_inz_frct
print(x, show_details = FALSE, ...)

Arguments

object

An inz_frct object representing the forecasts.

var

A character vector specifying the variable to summarize, or set to NULL to summarize all variables.

...

Additional arguments (ignored).

x

A `summary_inz_frct` object containing forecast summaries.

show_details

Logical; set to `TRUE` to show model details only when `pred_model` is an "ARIMA" model.

Value

A summary_inz_frct object containing the first few forecast observations, the forecasting model used, and its details (such as call, coefficients, and goodness of fit statistics).

See also

Examples

ts <- inzightts(visitorsQ, var = 2:5)
p <- predict(ts, "Japan")
s <- summary(p, "Japan")
s
#> 
#> 95% Prediction Interval
#> # A tsibble: 8 x 4 [1Q]
#>      Time Fitted Lower Upper
#>     <qtr>  <dbl> <dbl> <dbl>
#> 1 2012 Q2  2106. 1267. 2945.
#> 2 2012 Q3  3239. 2052. 4425.
#> 3 2012 Q4  2789. 1336. 4242.
#> 4 2013 Q1  4017. 2340. 5695.
#> 5 2013 Q2  1719. -397. 3835.
#> 6 2013 Q3  2851.  374. 5329.
#> 7 2013 Q4  2401. -391. 5194.
#> 8 2014 Q1  3630.  554. 6706.
#> 
#> Model:
#> # A mable: 1 x 1
#>                      Japan
#>                    <model>
#> 1 <ARIMA(0,1,0)(0,1,1)[4]>
#> 
print(s, show_details = TRUE)
#> 
#> 95% Prediction Interval
#> # A tsibble: 8 x 4 [1Q]
#>      Time Fitted Lower Upper
#>     <qtr>  <dbl> <dbl> <dbl>
#> 1 2012 Q2  2106. 1267. 2945.
#> 2 2012 Q3  3239. 2052. 4425.
#> 3 2012 Q4  2789. 1336. 4242.
#> 4 2013 Q1  4017. 2340. 5695.
#> 5 2013 Q2  1719. -397. 3835.
#> 6 2013 Q3  2851.  374. 5329.
#> 7 2013 Q4  2401. -391. 5194.
#> 8 2014 Q1  3630.  554. 6706.
#> 
#> Model:
#> # A mable: 1 x 1
#>                      Japan
#>                    <model>
#> 1 <ARIMA(0,1,0)(0,1,1)[4]>
#> 
#> 
#> Call:
#> .f(x = ..1, order = ..2, seasonal = ..3, xreg = ..4, include.mean = FALSE, fixed = ..7, 
#>     method = ..5)
#> 
#> Coefficients:
#>          sma1
#>       -0.4637
#> s.e.   0.1198
#> 
#> sigma^2 estimated as 183220:  log likelihood = -366.41,  aic = 736.82