Seasonal Subseries Plots for inzightts

Time series subseries plot by seasonal period

Usage

subseries(
  x,
  var = NULL,
  show_mean = TRUE,
  xlab = NULL,
  ylab = NULL,
  title = NULL
)

Arguments

x

A time series object represented by an inz_ts or tbl_ts object.

var

A character vector specifying the variable(s) to be plotted, or set to NULL to plot all variables.

show_mean

Logical; set to FALSE to exclude the mean line from the plot.

xlab

A title for the x-axis of the plot.

ylab

A title for the y-axis of the plot.

title

A title for the graph.

Value

A ggplot object of the seasonal subseries plot.

Details

Plots seasonal subseries of a time series represented by an inz_ts or tbl_ts object. Each subseries represents one seasonal period.

See also

Examples

t <- inzightts(visitorsQ)
if (FALSE) { # \dontrun{
subseries(t)
} # }